wasm_interp: move test_utils into tests/mod.rs
