+++ crates/wasm_interp/src/tests/test_f32.rs
@@ -1,6 +1,6 @@
 #![cfg(test)]
 
-use crate::test_utils::test_op_example;
+use super::test_op_example;
 use roc_wasm_module::{opcodes::OpCode, opcodes::OpCode::*, Value};
 
 fn test_f32_comparison(op: OpCode, arg1: f32, arg2: f32, expected: bool) {
