+++ crates/wasm_interp/src/tests/test_i32.rs
@@ -1,6 +1,6 @@
 #![cfg(test)]
 
-use crate::test_utils::test_op_example;
+use super::test_op_example;
 use roc_wasm_module::{opcodes::OpCode, opcodes::OpCode::*, Value};
 
 fn test_i32_binop(op: OpCode, arg1: i32, arg2: i32, expected: i32) {
